applied changes from 22457095fd36ea02a44652d45b5e7a6788cdea06              through 40accd16c45a4195e207b91c1d9dec6d6a8170cb
Improvements to the xdelta code in order to take advantage of the fact
that we are always using a relatively small window for the adler32
hash, and that when we skip forward, we normally skip forward by alot
so it's faster to just recompute the rolling checksum on the new data
than actually move the rolling checksum forward. 1.03x improvement in
cpu usage on the client, 1.12x improvement in cpu usage on the
server. This patch may actually be even better because I realized
during the development of the unit tests that the skip forward
checksum is always valid and should be applied if it is cheaper to
skip forward rather than incrementally move the checksum forward.
That said, the number of times that the attached patch would skip
forward is only slightly higher than the previous one as it is rare
for the window to move forward by more than 1 but less than the
checksum window.

adler32.hh: move a bunch of things into being constants, and migrate
to just doing a direct cast of the character rather than the widen()
stuff because the widen does a bunch of unnecessary masking.  Code
added to xdelta.cc to verify correctness of this transform.  Optimize
the insertion of a bunch of values for the case where the count is
small; just put an invariant in place to require that condition as
it's the only one used.  For the small update case, we can skip a lot
of the masking that we would otherwise have to do as nothing can
overflow.

xdelta.cc: mark the blocksz as constant (it is), remove the unused hi
variable from compute_delta_insns.  Specialize the xdelta advancement
code to handle the case where we advance by a lot by skipping over the
intermediate characters rather than pulling the rolling checksum
forward the entire way.  Add in two unit tests.  One verifies that the
compiler is sane, and the other verifies that the incremental and skip
forward checksums are equivalent in their calculations.