First pass at making mtn serve work with local domain sockets (also known as unix domain sockets).  Add a new option to hold the path of the socket.
Add code in the netsync.cc:serve_connections() to read this option and
construct a local uri which we can ask Netxx to bind to.

Modify the serve command to accept the new option.

Tangentially, fix what looks like a harmless bug in the options - many of
bind options had the wrong options group.