This monotone has Super Cow Powers.
Rosters in memory are now copy-on-write, and pulling auroraux or monotone
is about 16% faster. The roster cache doesn't know about this yet, so it
will horribly overestimate memory usage and evict rosters early.
roster_t::get_node() is dangerous now, it should return a const_node_t and
there should be a get_node_for_write() that will automatically unshare the
node before returning it.