Simplification of colorization purposes
The end game is to provide the user a simple map of purposes to a
given output color - via get_output_color. To help this, different
commands in Monotone may map to a single purpose such as files
being added (mtn status), and lines being added (mtn diff) both
represent an addition - so they are now treated as such.

In passing, also sorted the purposes alphabetically wherever they
are used for readability, with the exception for normal+reset as
these are special cases and should stand alone to reflect that.