Always print include/exclude patterns during a netsync operation
This is to inform the user which patterns are being used for the following
reasons

-> They could not specify any patterns explicitly on the command line making
any defaults in the db be used instead. As we print out "connecting to ..."
which could also be a default, this change helps the user also know what the
patterns are.

-> They could specify the patterns on the command line, and believe they
specified a pattern which is not actually being used. For instance, exclude
patterns require ;-, so the user enters

mtn pull mtn://myhost.com?myproject?good_branch;-bad_branch

However, this would actually do
connecting to mtn://myhost.com
	with include pattern = good_branch
	with exclude pattern = <blank>

then the terminal may throw an error after the netsync operation, as some
terminals require the ; be entered with \;. However, the user may believe
the pull succeeded and there is different error when we do not print the
patterns.

Us printing out the patterns here helps the user know exactly what has been
extracted from his URI, as not everyone knows or remembers to type \ when it
is needed.

The correct entry should be

mtn pull mtn://myhost.com?myproject?good_branch\;-bad_branch

-> I believe get_netsync_connect_command can also be used to specify patterns
so we should print them if they do not enter them on the command line.

applied changes from a92afea09d497594da69317ef9c661807ef38b2a
             through 9f4f219ea85d2f27ea603c8fd54d7928991ddbe7