Update from Botan2 to Botan3.
These are the pure changes made by Devstral2:123B with a few iterations, some of which I will try to factor out.

Prompts used (I don't remember exactly):
1. this is a project depending on obsolete Botan 2 library, please update it to Botan 3
2. the code does not work, use "make" and solve the problems
3. now it build but it has runtime problems, use "./mtn st" and solve the problems

This experiment took ~10 minutes of my time over a span of a few hours and a lot of tokens.

This is the tally (if it wasn't free right now):
Input tokens:  22 858 424 * 0.40$
Output tokens: 190 992 * 2$
Total:         9.52$